Refactor: Refactor arrow file into seperate three component files
